Found issue causing fatal error at end of import
When using threaded import we have to be careful to close the DB connection before any other tasks in each thread.
This is required as the DB resource handle is no longer valid after the fork

Additionally, the parent process has to wait on all of the child processes to complete and this causes the db connection
of the parent to be closed due to inactivity and a short timer on the DB server.  To solve this problem the connection
in the parent should be closed right after all the children have finished to ensure a fresh connection is opened.

id:19661